+++ misskey/src/lib.rs
@@ -21,8 +21,8 @@
 //!
 //! # #[tokio::main]
 //! # async fn main() -> anyhow::Result<()> {
-//! let client = HttpClient::builder("https://your.instance.example/api/".parse()?)
-//!     .token("API_TOKEN".to_string())
+//! let client = HttpClient::builder("https://your.instance.example/api/")
+//!     .token("API_TOKEN")
 //!     .build()?;
 //!
 //! client.create_note("Hello, Misskey").await?;
@@ -40,7 +40,7 @@
 //!
 //! # #[tokio::main]
 //! # async fn main() -> anyhow::Result<()> {
-//! let client = WebSocketClient::builder("wss://your.instance.example/streaming".parse()?)
+//! let client = WebSocketClient::builder("wss://your.instance.example/streaming")
 //!     .token("YOUR_API_TOKEN")
 //!     .connect()
 //!     .await?;
@@ -147,7 +147,7 @@ pub mod endpoint {
     //! # use misskey::{Client, HttpClient};
     //! # #[tokio::main]
     //! # async fn main() -> anyhow::Result<()> {
-    //! # let client = HttpClient::new("http://your.instance.example/api/".parse()?, Some("API_TOKEN".to_string()))?;
+    //! # let client = HttpClient::with_token("http://your.instance.example/api/".parse()?, "API_TOKEN")?;
     //! client
     //!     .request(
     //!         // Each endpoint implementation has a corresponding `Request` type.
@@ -175,7 +175,7 @@ pub mod endpoint {
     //! # use misskey::{Client, HttpClient};
     //! # #[tokio::main]
     //! # async fn main() -> anyhow::Result<()> {
-    //! # let client = HttpClient::new("http://your.instance.example/api/".parse()?, Some("API_TOKEN".to_string()))?;
+    //! # let client = HttpClient::with_token("http://your.instance.example/api/".parse()?, "API_TOKEN")?;
     //!     let me = client
     //!         .request(misskey::endpoint::i::Request::default())
     //!         .await?
@@ -205,7 +205,7 @@ pub mod streaming {
     //! # use misskey::WebSocketClient;
     //! # #[tokio::main]
     //! # async fn main() -> anyhow::Result<()> {
-    //! # let client = WebSocketClient::builder("ws://your.instance.example/streaming".parse()?)
+    //! # let client = WebSocketClient::builder("wss://your.instance.example/streaming")
     //! #     .token("API_TOKEN")
     //! #     .connect()
     //! #     .await?;
@@ -231,7 +231,7 @@ pub mod streaming {
     //! # use misskey::WebSocketClient;
     //! # #[tokio::main]
     //! # async fn main() -> anyhow::Result<()> {
-    //! # let client = WebSocketClient::builder("ws://your.instance.example/streaming".parse()?)
+    //! # let client = WebSocketClient::builder("wss://your.instance.example/streaming")
     //! #     .token("API_TOKEN")
     //! #     .connect()
     //! #     .await?;
@@ -269,7 +269,7 @@ pub mod streaming {
     //! # use misskey::WebSocketClient;
     //! # #[tokio::main]
     //! # async fn main() -> anyhow::Result<()> {
-    //! # let client = WebSocketClient::builder("ws://your.instance.example/streaming".parse()?)
+    //! # let client = WebSocketClient::builder("wss://your.instance.example/streaming")
     //! #     .token("API_TOKEN")
     //! #     .connect()
     //! #     .await?;