+++ misskey-websocket/CHANGELOG.md
@@ -16,3 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Follow changes of `Client` in `misskey-core`
   - Adjust lifetime specification of request methods
 - Implement `StreamingClient`
+- Improve API of `ReconnectCondition` and `ReconnectConfig`
+- Improve API of `WebSocketClient`
+- Improve API of `WebSocketClientBuilder`